refactor: split off files from hook.go
+ move loader tests
